test: extract byText related tests
